requirements.txt
websocket-client==1.9.0
requests==2.31.0
